pyproject.toml
[tool.poetry]
name = "flask-masterclass"
version = "0.1.0"
description = ""
authors = ["Wesley Júnior <agarwesley19@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dynaconf = "^3.1.7"
urllib3 = "^1.26.8"
mvc-flask = "^2.3.1"
Flask-SQLAlchemy = "^2.5.1"
black = "^22.1.0"
Flask-Migrate = "^3.1.0"

[tool.poetry.dev-dependencies]
ward = "^0.65.0-beta.0"
splinter = {version = "^0.17.0", extras = ["flask"]}
python-dotenv = "^0.19.2"
selenium = "^4.1.3"
Faker = "^13.3.1"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
